refactor(product-ui): split shared chat composer surfaces
