+++ apps/packages/product-domain/src/chats/cloud/composer-launch-defaults.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_DIRECT_PROMPT_MODEL_ID = "us.anthropic.claude-sonnet-4-6";
+export const DEFAULT_DIRECT_PROMPT_AGENT_KIND = "claude";
